requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.57.0
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
packaging==25.0
pillow==10.4.0
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.17.0
zipp==3.20.2
